new mechanisms to automatically read codemeta.json
Instead of updating the main-list/projects.json, one only needs to
provide the URL of the codemeta.json file for their softwares.

It will be automatically read and will update the main-list/projects.json
if needed.

This code does not respect the grammar of main-list/projects.json as
it defines the spack and guix packages in a unique package field.

This does not completely work yet, as the main-list/projects.json
needs the fields documentation and packages which are not yet present
in the codemeta grammar as defined in https://doi.org/10.5063/schema/codemeta-2.0

We would either need to extend the grammar or define a specific numpex
grammar.

Hence the example for StarPU is using a codemeta.json file which is
not valid as it defines the documentation and package fields.